Update train.py
Support user-defined point selection for concepts, concept visualization, and selection.  Successfully integrate the combined image of two concepts into the program, and solve the problem that the program does not support the prompt such as "v1 and v2", so that successfully add the epsilon MSE loss module and cross-attention regularization module to the program. The token average is taken is taken in the test generation (inferring) module for comparison.